pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "task_manager"
readme = "README.md"
requires-python = ">=3.13"
packages = [
  { include = "task_manager" },
]
dependencies = [
  "Django>=5.0",
  "gunicorn",
  "python-dotenv",
  "dj-database-url",
  "psycopg2-binary",
  "django-bootstrap5>=25.1",
  "uv",
  "django-filter>=25.1",
]

[project.optional-dependencies]
filters = ["django-filter"]
